# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restobar Venecia Peruana",
        "description": "Sistema de restaurante en linea. Con control de inventario",
        "contact": {
            "email": "neisserrey@upeu.edu.pe"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://veneciaperuana.dev.com/api/",
            "description": "Documentación del Restobar Venecia Peruana"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Mostrar Productos",
                "operationId": "App\\Http\\Controllers\\Api\\ProductController::index",
                "responses": {
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Lista de Productos. Correcto"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Agregar un nuevo Producto.",
                "operationId": "store",
                "requestBody": {
                    "description": "Producto a ser creado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "201": {
                        "description": "Creado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/products/{productId}": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Buscar Producto por ID",
                "description": "Retorna un Producto",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID del Producto a retornar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "404": {
                        "description": "Producto no existe"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "products"
                ],
                "summary": "Actualizar un Producto existente",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID del Producto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Producto a ser actualizada",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Producto no encontrada"
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Producto actualizada correctamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "products"
                ],
                "summary": "Elimina un Producto",
                "description": "",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "Id del Producto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Producto eliminado correctamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "description": "Name of the Product",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description short of the Product",
                        "type": "string"
                    },
                    "code": {
                        "description": "Code of the Product",
                        "type": "string"
                    },
                    "category_id": {
                        "description": "The id of the category of the Product",
                        "type": "int"
                    },
                    "created_at": {
                        "description": "When the Product was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "When the Product was updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}